% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biologicalReps.R
\name{biologicalReps}
\alias{biologicalReps}
\title{calculates all of the biological replicate means}
\usage{
biologicalReps(x)
}
\arguments{
\item{x}{A list generated with technicalReps}
}
\description{
calculates all of the biological replicate means
}
