% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPlate.R
\name{checkPlate}
\alias{checkPlate}
\title{Create a snapshot of a specific time point for a plate}
\usage{
checkPlate(input_file, variable, time_point, log_scale = TRUE)
}
\arguments{
\item{input_file}{A Victor X3 output file (in format .txt)}

\item{variable}{A variable for visualization}

\item{time_point}{Which of the time points (Repeat) to visualize?}

\item{log_scale}{(logical) Should the output plot be log-scaled?}
}
\description{
Create a snapshot of a specific time point for a plate
}
