% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkReplicates.R
\name{checkReplicates}
\alias{checkReplicates}
\title{Visualize data points for specific Wells over time}
\usage{
checkReplicates(input, sample, measure = "Absorbance")
}
\arguments{
\item{input}{A list generated by readPlate()}

\item{sample}{A Character string for the sample you'd like to see (either a sample name or #blank)}

\item{measure}{A character string (or part of it) that identifies the measurement of interest (e.g. "Absorbance" or "Abs", "GFP")}
}
\description{
Visualize data points for specific Wells over time
}
