% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseTechnicalReps.R
\name{collapseTechnicalReps}
\alias{collapseTechnicalReps}
\title{Calculate technical replicate means for a single sample}
\usage{
collapseTechnicalReps(
  sample_name,
  plate,
  sample_wells,
  blank_wells,
  normalizeFluorescence = FALSE,
  ignoreBlanks = FALSE
)
}
\arguments{
\item{sample_name}{The name for the sample to be collapsed}

\item{plate}{The plate rawdata from which to calculate means}

\item{sample_wells}{A character vector pointing the coordinates (e.g. "A01", "B02"...) for the sample within the plate}

\item{blank_wells}{A character vector pointing the coordinates for the wells of the 'blank'}

\item{normalizeFluorescence}{(logical) Should the fluorescence values be normalized by the absorbance?}

\item{ignoreBlanks}{If TRUE, the blank wells mean is not subtracted from the sample means}
}
\value{
A small tibble with the calculated mean for the technical replicates
}
\description{
Calculate technical replicate means for a single sample
}
