% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLayout.R
\name{createLayout}
\alias{createLayout}
\title{Generate a 96-well map template}
\usage{
createLayout(layout_name, interactive = FALSE)
}
\arguments{
\item{layout_name}{A string specifying the name for the map template.}

\item{interactive}{Use interactive mode to design template}
}
\description{
Generate a 96-well map template
}
