% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMap.R
\name{createMap}
\alias{createMap}
\title{Generate a 96-well map template}
\usage{
createMap(map_name, interactive = FALSE)
}
\arguments{
\item{map_name}{A string specifying the name for the map template.}

\item{interactive}{Use interactive mode to design template}
}
\description{
Generate a 96-well map template
}
