% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeWideTable.R
\name{makeWideTable}
\alias{makeWideTable}
\title{Converts a biologicalRep table from long to wide}
\usage{
makeWideTable(bioReps)
}
\arguments{
\item{bioReps}{a table generated by biologicalReps()}
}
\description{
Converts a biologicalRep table from long to wide
}
