% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergePlates.R
\name{mergePlates}
\alias{mergePlates}
\title{calculates all of the biological replicate means}
\usage{
mergePlates(inputlist)
}
\arguments{
\item{inputlist}{A vector of list names to be collapsed}
}
\description{
calculates all of the biological replicate means
}
