% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nivoParser.R
\name{nivoParser}
\alias{nivoParser}
\title{Parse Victor Nivo long output file format}
\usage{
nivoParser(input_file, delimiter = "\\t", point_measure = FALSE)
}
\arguments{
\item{input_file}{A microplate reader output file (in formats .txt or .csv)}

\item{delimiter}{Field delimiter used in the file}

\item{point_measure}{for mode = "nivo", tells whether the output file is a point measurement, or if there are several points into the same file}
}
\description{
Parse Victor Nivo long output file format
}
