% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateParser.R
\name{plateParser}
\alias{plateParser}
\title{reads a plate map from a file}
\usage{
plateParser(plate_file, plate_map)
}
\arguments{
\item{plate_file}{A VictorX output file}

\item{plate_map}{A .xlsx spreadsheet mapping samples to wells in a 96 well plate}
}
\value{
a list containing the formatted table map and metadata, as well as the raw data from plate_file
}
\description{
reads a plate map from a file
}
