% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPlateInput.R
\name{readPlateInput}
\alias{readPlateInput}
\title{Read plate input file}
\usage{
readPlateInput(input_file)
}
\arguments{
\item{input_file}{A Victor X3 output file (in format .txt)}
}
\value{
A tibble
}
\description{
Read plate input file
}
