% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technicalReps.R
\name{technicalReps}
\alias{technicalReps}
\title{calculates all of the technical replicate means}
\usage{
technicalReps(plate_file, plate_map)
}
\arguments{
\item{plate_file}{A VictorX output file}

\item{plate_map}{A .xlsx spreadsheet mapping samples to wells in a 96 well plate}
}
\description{
calculates all of the technical replicate means
}
