% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softmaxParser.R
\name{timeToRepeats}
\alias{timeToRepeats}
\title{Parse SoftMax's plate reader software long data format \link{EXPERIMENTAL}}
\usage{
timeToRepeats(string)
}
\arguments{
\item{input_file}{A  SoftMax reader output file (when exproting, select long column format)}
}
\value{
A tibble
}
\description{
Parse SoftMax's plate reader software long data format \link{EXPERIMENTAL}
}
