% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPlate.R
\name{addPlate}
\alias{addPlate}
\title{calculates all of the technical replicate means}
\usage{
addPlate(file, layout)
}
\arguments{
\item{file}{A VictorX output file}

\item{layout}{A .xlsx spreadsheet mapping samples to wells in a 96 well plate}
}
\description{
calculates all of the technical replicate means
}
