% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biologicalReps.R
\name{biologicalReps}
\alias{biologicalReps}
\title{calculates all of the biological replicate means}
\usage{
biologicalReps(inputlist)
}
\arguments{
\item{inputlist}{A vector of list names to be collapsed}

\item{normalizeFluorescence}{(logical) Should the fluorescence values be normalized by the absorbance?}
}
\description{
calculates all of the biological replicate means
}
