% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateParser.R
\name{plateParser}
\alias{plateParser}
\title{reads a plate plate layout from a file}
\usage{
plateParser(
  file,
  layout,
  mode = "x3",
  delimiter = "\\t",
  point_measure = FALSE
)
}
\arguments{
\item{file}{A VictorX output file}

\item{layout}{A .xlsx spreadsheet mapping samples to wells in a 96 well plate}

\item{mode}{Which parsing function to use. Currently supports "x3" for Victor X3 conventional output file, or "nivo" for Victor Nivo files}

\item{delimiter}{Field delimiter used in the file}

\item{point_measure}{for mode = "nivo", tells whether the output file is a point measurement, or if there are several points into the same file}
}
\value{
a list containing the formatted table map and metadata, as well as the raw data from plate_file
}
\description{
reads a plate plate layout from a file
}
